Turn create event into pop up modal. Place the map in landing page
