+++ components/createEventModal/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Button, Form, Modal } from "react-bootstrap";
+import { Button, Form, Modal, Col, Row } from "react-bootstrap";
 
 const CreateEventModal = ({
   showModal,
@@ -8,27 +8,22 @@ const CreateEventModal = ({
   showModal: boolean;
   setShowModal: Function;
 }) => {
-  const eventTitle = "";
-  const eventCategory = "";
-  const eventDescription = "";
-  const eventDatetime = "";
-  const eventHostInfo = "";
-
-  const [eventData, setEventData] = useState({
-    eventTitle,
-    eventCategory,
-    eventDescription,
-    eventDatetime,
-    eventHostInfo,
-  });
-
-  const updateInputs = (form: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = form.target;
-    setEventData({ ...eventData, [name]: value });
-  };
+  const [title, setTitle] = useState("");
+  const [description, setDescription] = useState("");
+  const [date, setDate] = useState("");
+  const [category, setCategory] = useState("");
+  const [address, setAddress] = useState("");
+  const [postcode, setPostCode] = useState("");
+  const [validated, setValidated] = useState(false);
 
   // TODO: edit this function to send eventData to backend
-  const handleSubmit = () => {
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    const form = event.currentTarget;
+    if (form.checkValidity() === false) {
+      event.preventDefault();
+      event.stopPropagation();
+    }
+    setValidated(true);
     console.log("Submit button clicked, please pass the data to backend");
     setShowModal(false);
   };
@@ -36,66 +31,117 @@ const CreateEventModal = ({
   return (
     <Modal show={showModal}>
       <Modal.Body>
-        <Form noValidate onSubmit={handleSubmit}>
+        <Form noValidate validated={validated} onSubmit={handleSubmit}>
           <p style={{ fontSize: "larger", fontWeight: "bolder" }}>
             Create an event
           </p>
 
-          <Form.Group>
-            <Form.Label>Event Title</Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Event Title"
-              name="eventTitle"
-              onChange={updateInputs}
-              required
-            />
+          <Form.Group as={Row} className="general-form-group">
+            <Form.Label column sm={2}>
+              Title
+            </Form.Label>
+            <Col>
+              <Form.Control
+                required
+                value={title}
+                onChange={(event) => setTitle(event.target.value)}
+                placeholder="Event title"
+              />
+              <Form.Control.Feedback type="invalid">
+                Event title required
+              </Form.Control.Feedback>
+            </Col>
           </Form.Group>
 
-          <Form.Group>
-            <Form.Label>Event Category</Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Event Category"
-              onChange={updateInputs}
-              required
-              name="eventCategory"
-            />
+          <Form.Group as={Row} className="general-form-group">
+            <Form.Label column sm={2}>
+              Date
+            </Form.Label>
+            <Col>
+              <Form.Control
+                required
+                type="date"
+                value={date}
+                onChange={(event) => setDate(event.target.value)}
+              />
+            </Col>
           </Form.Group>
 
-          <Form.Group>
-            <Form.Label>Event Description</Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Event Description"
-              onChange={updateInputs}
-              required
-              name="eventDescription"
-            />
+          <Form.Group as={Row} className="general-form-group">
+            <Form.Label column sm={2}>
+              Category
+            </Form.Label>
+            <Col>
+              <Form.Control
+                required
+                as="select"
+                value={category}
+                onChange={(event) => setCategory(event.target.value)}
+              >
+                <option>Volunteering</option>
+                <option>Community Event</option>
+                <option>Local Meetup</option>
+              </Form.Control>
+            </Col>
           </Form.Group>
+          <Row>
+            <Col>
+              <Form.Group as={Row} className="general-form-group">
+                <Form.Label column sm={4}>
+                  Address
+                </Form.Label>
+                <Col sm={8}>
+                  <Form.Control
+                    required
+                    type="text"
+                    placeholder="Address"
+                    name="city"
+                    value={address}
+                    onChange={(event) => setAddress(event.target.value)}
+                  />
 
-          <Form.Group>
-            <Form.Label>Event Datetime</Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Event Datetime"
-              onChange={updateInputs}
-              required
-              name="eventDatetime"
-            />
+                  <Form.Control.Feedback type="invalid">
+                    Address required
+                  </Form.Control.Feedback>
+                </Col>
+              </Form.Group>
+            </Col>
+            <Col>
+              <Form.Group as={Row} className="general-form-group">
+                <Form.Label column sm={3}>
+                  Post Code
+                </Form.Label>
+                <Col>
+                  <Form.Control
+                    required
+                    type="text"
+                    placeholder="PostCode"
+                    value={postcode}
+                    onChange={(event) => setPostCode(event.target.value)}
+                  />
+
+                  <Form.Control.Feedback type="invalid">
+                    Required
+                  </Form.Control.Feedback>
+                </Col>
+              </Form.Group>
+            </Col>
+          </Row>
+
+          <Form.Group as={Row} className="general-form-group">
+            <Form.Label column sm={2}>
+              Description
+            </Form.Label>
+            <Col>
+              <Form.Control
+                required
+                as="textarea"
+                value={description}
+                rows={4}
+                onChange={(event) => setDescription(event.target.value)}
+              />
+            </Col>
           </Form.Group>
-
-          <Form.Group>
-            <Form.Label>Event Host</Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Event Host"
-              onChange={updateInputs}
-              required
-              name="eventHostInfo"
-            />
-          </Form.Group>
-
           <div style={{ marginTop: "20px" }}>
             <Button id="createButton" type="submit" style={{ margin: "5px" }}>
               Create Event
