--- styles/CreateEventForm.module.css
@@ -1,25 +1,0 @@
-
-.main {
-    min-height: 90vh;
-    padding: auto;
-    display: flex;
-    width: 50%;
-    flex-direction: column;
-    justify-content: center;
-  }
-
-.formGroup{
-    width: 100%;
-    font-size: 1.5rem;
-
-}
-
-.formControl{
-    font-size: 2rem;
-    margin-left: 1rem;
-}
-
-.button{
-    display: flex;
-    justify-content: center;
-}
