+++ styles/globals.css
@@ -32,5 +32,7 @@ a {
 }
 
 .general-form-group {
+  padding-left: 5px;
+  padding-right: 5px;
   padding-top: 10px;
 }
